% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsebird-package.R
\docType{package}
\name{bbsebird}
\alias{bbsebird}
\alias{bbsebird-package}
\title{bbsebird: Unifying BBS and eBird Data for Integrating Modeling}
\description{
Imports, munges and aligns eBird and BBS data to a spatial sampling grid of user's select size. Additionally, produces bundles of data for use in JAGS, Nimble or STAN and runs select BUGS models in parallel in JAGS and Nimble via R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trashbirdecology.github.io/bbsebird/}
}

}
\author{
\strong{Maintainer}: Jessica Burnett \email{jburnett@usgs.gov} (\href{https://orcid.org/0000-0002-0896-5099}{ORCID})

Other contributors:
\itemize{
  \item Paige Howell (\href{https://orcid.org/0000-0002-5911-2870}{ORCID}) [contributor]
  \item Andy Royle (\href{https://orcid.org/0000-0003-3135-2167}{ORCID}) [contributor]
}

}
\keyword{internal}
