% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data.r
\name{bundle_long}
\alias{bundle_long}
\title{Bundle Data for Use in JAGS (LONG format)}
\usage{
bundle_long(bbs_spatial, ebird_spatial, grid)
}
\description{
An updated version of bundle_data, where site by year elements are provided in long format
}
