% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_cols}
\alias{convert_cols}
\title{Convert columns}
\usage{
convert_cols(x)
}
\arguments{
\item{x}{The data frame with one or more columns.}
}
\description{
Converts columns classes based on names for the bbs and ebird data
}
\keyword{internal}
