% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_spec}
\alias{dir_spec}
\title{Specify input and output directories.
Produces a list comprising directories for munged data, JAGS data lists, figures, and model outputs. Should be used with \code{list2env()} to assign directories to desired environment (typically .GlobalEnv)}
\usage{
dir_spec(dir.orig.data)
}
\arguments{
\item{dir.orig.data}{Location of the original BBS and eBird data. This directory should house multiple directories, including the BBS route shapefiles, the eBird database.}
}
\description{
Specify input and output directories.
Produces a list comprising directories for munged data, JAGS data lists, figures, and model outputs. Should be used with \code{list2env()} to assign directories to desired environment (typically .GlobalEnv)
}
