% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{do_indexing}
\alias{do_indexing}
\title{Make 2-D array for Indexing Purposes}
\usage{
do_indexing(X)
}
\arguments{
\item{X}{a data frame with variables site.id, site.ind, grid.id, grid.ind, year.id, year.ind, and "c"}
}
\description{
Make 2-D array for Indexing Purposes
}
\keyword{internal}
