% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubcorms-package.R
\docType{package}
\name{dubcorms-package}
\alias{dubcorms}
\alias{dubcorms-package}
\title{dubcorms: Unifying BBS and eBird Data}
\description{
Imports, munges and integrates eBird and BBS data and applies them to a spatial sampling grid. This package also outputs data in various forms for use in JAGS, {\codemcgv::jagam()}
}
\author{
\strong{Maintainer}: Jessica Burnett \email{jburnett@usgs.gov} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
