% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bbs_by_species.r
\name{filter_bbs_by_species}
\alias{filter_bbs_by_species}
\title{Subset BBS observations data by one or more species of interest.}
\usage{
filter_bbs_by_species(
  list,
  search = "Double-crested Cormorant",
  zero.fill = TRUE
)
}
\arguments{
\item{list}{A list with element "species_list", obtained from running bbsAssistant::get_bbs_data()...}

\item{search}{A vector of one or more species (using English Common Name) to subset the data by. Capitalization ignored.}

\item{zero.fill}{If TRUE and a single species is provided in 'search', this function will output list$observations with zero-filled data.}
}
\description{
...
}
