% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ebird_data.R
\name{filter_ebird_data}
\alias{filter_ebird_data}
\title{Create and Write or Load In the Filtered eBird Data}
\usage{
filter_ebird_data(
  fns.ebird,
  dir.ebird.out,
  countries = NULL,
  states = NULL,
  complete.only = TRUE,
  protocol = c("Traveling", "Stationary"),
  species = "Double-crested Cormorant",
  overwrite = FALSE,
  remove.bbs.obs = TRUE
)
}
\arguments{
\item{fns.ebird}{File paths for the EBD and SamplingEvents data to import. Character vector of filenames for original files.}

\item{dir.ebird.out}{Location of where to save and find the filtered/subsetted data.
If not specified will default to subdir in project directory.}

\item{overwrite}{Logical. If true will overwrite existing filtered data objects in project directory.}
}
\description{
Filter the eBird data and sampling events using R package AUK.
}
