% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-longlat_to_utm.R
\name{find_UTM_zone}
\alias{find_UTM_zone}
\title{Functions for converting longitude and latitude to UTM zones and coordinates.}
\source{
https://stackoverflow.com/questions/18639967/converting-latitude-and-longitude-points-to-utm
}
\usage{
find_UTM_zone(longitude, latitude)
}
\arguments{
\item{longitude}{vector of X coordinate(s)}

\item{latitude}{vector of Y coordinate(s)}
}
\description{
Functions for converting longitude and latitude to UTM zones and coordinates.
}
