% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_list_dims.r
\name{get.list.dims}
\alias{get.list.dims}
\title{Get Dimensions of List Elements}
\usage{
get.list.dims(list, print = TRUE)
}
\arguments{
\item{list}{List of elements}

\item{print}{logical if TRUE will print list element dimensions to screen.}
}
\description{
Get Dimensions of List Elements
}
