% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebird_obs_files.R
\name{get_ebird_obs_files}
\alias{get_ebird_obs_files}
\title{Identify and Unpack eBird Observations Files}
\usage{
get_ebird_obs_files(
  dir.ebird.in,
  mmyyyy = "feb-2022",
  dir.out = NULL,
  species = NULL,
  countries = NULL
)
}
\arguments{
\item{dir.out}{first, search for existing .rds files in output directory.}
}
\description{
Identify and Unpack eBird Observations Files
}
