% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ebird_sampling_events.R
\name{get_ebird_sampling_events}
\alias{get_ebird_sampling_events}
\title{Retrieve the eBird sampling events from local directory.}
\usage{
get_ebird_sampling_events(
  f_samp = NULL,
  dir = NULL,
  dir.out = "data-local/ebird/sampling-out/",
  n = 8e+06,
  n.amer = TRUE,
  complete = TRUE,
  protocol.type = c("Traveling", "Stationary")
)
}
\arguments{
\item{f_samp}{Filename for the sampling .txt file. Either f_samp or dir must be specified.}

\item{dir}{Directory for where the sampling event file(s) are stored. Either f_samp or dir must be specified.}

\item{dir.out}{Directory for where to store the subsets of the sampling events data. If not specified will default to data-local/ebird.}

\item{n.amer}{Logical. If working with North American countries, significantly reduces the file sizes of the sampling events data.}

\item{overwrite}{Logical. If .RDS file, "sampling_events.rds" exists in dir or dir.out, will NOT overwrite when FALSE.}
}
\description{
The eBird sampling events data is quite large, often maxing out available RAM even when R's memory.limit() is set to max. This function splits the sampling event data into smaller files and uses \code{auk} functions to import.
}
