% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zerofilled_ebird.R
\name{get_zerofilled_ebird}
\alias{get_zerofilled_ebird}
\title{Create the Zero-filled eBird Data}
\usage{
get_zerofilled_ebird(
  fns = NULL,
  mmyyyy = "oct-2021",
  species = NULL,
  overwrite = FALSE,
  dir.ebird.munged = "data-local/ebird/"
)
}
\arguments{
\item{fns}{File paths for the EBD and SamplinEvents data to import. Character vector of filenames for original files.}

\item{mmyyyy}{Month and year of desired input data as specified in fns. This format is default for eBird downloads.}

\item{species}{Names of species of interest.}

\item{overwrite}{Logical. If true,}

\item{dir.ebird.munged}{Location of where to save and find the filtered, zero-filled data. If not specified will default to subdir in project directory.}
}
\description{
Creates zero-filled data from select eBird EBD (ebd_rel) and Sampling Events (ebd_sampling) files. If the files already exist in dir.ebird.munged
}
