% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jdat.R
\name{import_jdat}
\alias{import_jdat}
\title{Import the Prepared RDS Data For Use in JAGS}
\usage{
import_jdat(dir.jags, filename = "jdat.rds")
}
\arguments{
\item{dir.jags}{Path name to where the jags data files (.rds) are stored}

\item{filename}{Name of the file to import.}
}
\description{
Import the Prepared RDS Data For Use in JAGS
}
