% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jdat.contents}
\alias{jdat.contents}
\title{Species Lookup Table for BBS and eBird Data}
\format{
A data frame
\describe{
\item{name}{object name}
\item{description}{informal description of item}
\item{note}{supplementary information}
\item{class}{expected class of object}
\item{dims}{expected dimensions of object, relative to parent list. Not provided for objects of class list or single}
}
}
\usage{
jdat.contents
}
\description{
\describe{
\item{}{}
}
}
\details{
Descriptions of JDAT Contents
}
\keyword{datasets}
