% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{junk_it}
\alias{junk_it}
\title{Keep Only Wanted Objects in Memory}
\usage{
junk_it(args.save, new.args.save = NULL)
}
\arguments{
\item{args.save}{A list of new or previously specified arguments.}

\item{new.args.save}{Optional. One or more new object names to save (in addition to args.save)}
}
\description{
Removes all items except those specified. Returns a new list of the arguments you want to always retain in memory while removing duplicates.
}
\keyword{internal}
