% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-longlat_to_utm.R
\name{longlat_to_UTM}
\alias{longlat_to_UTM}
\title{Convert long, lat Locations to UTM coordinates and zones
Returns a data frame containing the UTM values, the zone and the hemisphere associated with long lat coordinates.}
\usage{
longlat_to_UTM(long, lat, units = "m")
}
\arguments{
\item{long}{vector of X coordinates}

\item{lat}{vector of Y coordinates}

\item{units}{one of c('m', 'km') returns UTM coordinates in meters or kilometers, respectively.}
}
\description{
Convert long, lat Locations to UTM coordinates and zones
Returns a data frame containing the UTM values, the zone and the hemisphere associated with long lat coordinates.
}
