% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.model.data.R
\name{make.model.data}
\alias{make.model.data}
\title{Make List of Data and Constants for Nimble/JAGS}
\usage{
make.model.data(data)
}
\arguments{
\item{data}{List of data objects and constants as produced by make_bundle}
}
\description{
Make List of Data and Constants for Nimble/JAGS
}
