% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_array.R
\name{make_array}
\alias{make_array}
\title{Cast Data to Array or Matrix for Use in JAGS}
\usage{
make_array(
  df = bbs,
  val,
  row = c("rteno", "checklist_id"),
  col = "year",
  slice = "gridcellid",
  drop.na.rows = FALSE
)
}
\arguments{
\item{df}{Data frame with three or more columns}

\item{val}{target value to populate the matri(ces)}

\item{row}{target rownames}

\item{col}{target colnames}

\item{slice}{target slice-by values}

\item{drop.na.rows}{logical if TRUE will drop the values of arg "val" where is NA.}
}
\description{
Cast Data to Array or Matrix for Use in JAGS
}
