% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ebird.R
\name{make_ebird}
\alias{make_ebird}
\title{Munge eBird for Use Case}
\usage{
make_ebird(
  dir.ebird.in,
  dir.out = NULL,
  fns.samps = NULL,
  fns.obs = NULL,
  mmyyyy = "fEb-2022",
  overwrite = FALSE,
  countries = c("US", "CA"),
  states = NULL,
  species = NULL,
  complete.only = TRUE,
  protocol = c("Traveling", "Stationary"),
  remove.bbs.obs = TRUE,
  years = NULL,
  max.effort.km = NULL,
  max.effort.mins = NULL,
  max.num.observers = 10,
  ncores = NULL
)
}
\arguments{
\item{dir.ebird.in}{Where the original eBird data are stored. A single directory only...}

\item{dir.out}{Path to directory for where to save compressed outputs of munged data. If NULL will save to a subdirectory within current working directory.}

\item{fns.samps}{Filenames for import sampling events}

\item{fns.obs}{Filenames or directory for}

\item{mmyyyy}{month-year used for searching ebird files.}

\item{overwrite}{if TRUE will overwrite any existing files.}

\item{countries}{countries}
}
\description{
Munge eBird for Use Case
}
