% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ebird_spatial.R
\name{make_ebird_spatial}
\alias{make_ebird_spatial}
\title{Create an SF object for eBird Observations}
\usage{
make_ebird_spatial(x, crs.target, grid = NULL)
}
\arguments{
\item{x}{A data frame with columns c(latitude, longitude) and optional attributes (columns).}

\item{crs.target}{the integer representing the target CRS.}

\item{grid}{a spatial grid over which the eBird data will be overlaid.}
}
\description{
Converts eBird (preferably zero-filled data) into a simple features spatial object.
}
