% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gam.R
\name{make_gam}
\alias{make_gam}
\title{Function for Generating Basis Functions}
\usage{
make_gam(
  coords,
  method = "cubic2d",
  nd = 20,
  nruns = 10,
  num.nn = 20,
  print.plot = TRUE,
  max.loop = 40,
  K = NULL,
  scale.coords = FALSE
)
}
\arguments{
\item{method}{one of c("mgcv", "cubic2d", "strebel")}

\item{nd}{...}

\item{nruns}{...}

\item{num.nn}{...}

\item{print.plot}{logical if TRUE will print the spatial representation of gam to local device}

\item{max.loop}{...}

\item{K}{max number of basis functions to produce in mgcv::jagam()}

\item{scale.coords}{logical if TRUE will scale the XY coordinates}

\item{XY}{matrix or data frame comprising XY coordinates, where first column is X coordinate and second is Y.}
}
\description{
Function for Generating Basis Functions
}
