% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid_overlay.R
\name{make_grid_overlay}
\alias{make_grid_overlay}
\title{Make Spatial Grid Overlay}
\usage{
make_grid_overlay(extent = NULL, grid.type = "hexagonal", dir.grid.out = NULL)
}
\arguments{
\item{extent}{OBJ TYPE The extent of the spatial grid.}

\item{grid.type}{One of c("hexagonal", "square").}
}
\description{
Create a spatial grid overlay for assigning observation and covariate data to spatial point processes.
}
\keyword{internal}
