% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_indexing_df.R
\name{make_indexing_df}
\alias{make_indexing_df}
\title{Make 2-D array (Matrix for JAGS)}
\usage{
make_indexing_df(X, Y)
}
\arguments{
\item{X}{a data frame with dimensions site by year}

\item{Y}{a data frame with dimensions grid by site}
}
\description{
Make 2-D array (Matrix for JAGS)
}
\keyword{internal}
