% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-and-inits.R
\name{make_inits_list}
\alias{make_inits_list}
\title{Create a List of Repeated Initial Values for each Chain}
\usage{
make_inits_list(inits, nc = 3)
}
\arguments{
\item{inits}{a list of initial values which will be repeated according to arg \code{nc}}

\item{nc}{number of chains (number of times the inits will be repeated)}
}
\description{
Create a List of Repeated Initial Values for each Chain
}
