% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jags_list.R
\name{make_jags_list}
\alias{make_jags_list}
\title{Make a List for Use in JAGS}
\usage{
make_jags_list(input)
}
\arguments{
\item{input}{character vector names of data objects in the workspace to be included in the outputted list}
}
\description{
Make a List for Use in JAGS
}
