% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jags_list_old.R
\name{make_jags_list_old}
\alias{make_jags_list_old}
\title{Munge and Output a List or List of Lists for Use in JAGS}
\usage{
make_jags_list_old(
  dat,
  dat.names = NULL,
  drop.nas = FALSE,
  max.C.ebird = 100,
  scale.vars = TRUE,
  dir.out,
  dir.plots = NULL,
  fn.out = "jdat"
)
}
\arguments{
\item{dat}{a single data object or a list of multiple objects to munge and collapse into a list of data for use in JAGS}

\item{dat.names}{character string or vector of strings for names of input data objects. If not supplied, the function will guess whether the data is ebird, bbs, or grid.}

\item{max.C.ebird}{if TRUE will drop all checklists where number of observed birds is greater than 100. This is an arbitrary number.}

\item{scale.vars}{Logical If TRUE will scale variables. This needs to be improved/checked.}

\item{dir.out}{Directory within which the output list will be saved  as "jdat"}

\item{dir.plots}{where to save density plots of counts}

\item{fn.out}{Filename of the object output as a .RDS file}
}
\description{
Munge and Output a List or List of Lists for Use in JAGS
}
