% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_mat}
\alias{make_mat}
\title{Make 2-D array (Matrix for JAGS)}
\usage{
make_mat(df.in, row = "site.ind", col = "year.ind", val, replace.na = FALSE)
}
\arguments{
\item{df.in}{Data frame}

\item{row}{Variable in df containing target row names}

\item{col}{Variable in df containing target column names}

\item{val}{Variable containing the target cell contents}

\item{replace.na}{Whether to replace NA values with zero.}
}
\description{
Make 2-D array (Matrix for JAGS)
}
