% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_data.R
\name{make_model_data}
\alias{make_model_data}
\title{Make List of Data and Constants for Nimble/JAGS}
\usage{
make_model_data(data)
}
\arguments{
\item{data}{List of data objects and constants as produced by make_bundle}
}
\description{
Make List of Data and Constants for Nimble/JAGS
}
