% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_grid.R
\name{make_spatial_grid}
\alias{make_spatial_grid}
\title{Create a Spatial Sampling Grid}
\usage{
make_spatial_grid(
  dir.out,
  overwrite = TRUE,
  countries = c("US", "CA"),
  states = NULL,
  map.view = c("static", "interactive"),
  crs.target = 4326,
  hexagonal = TRUE
)
}
\arguments{
\item{dir.out}{Where to save the resulting grid (as .RDS).}

\item{overwrite}{TRUE will overwrite existing \code{grid.rds} in \code{dir.out}}

\item{countries}{Vector of countries. Defaults to a base map of USA and CAN, unless arg \code{states} is provided. If arg \code{states} is provided, this argument will be ignored. Must be specified using ISO-A2}

\item{states}{Vector of states to which the spatial grid will be clipped. Must be specified using ISO 3166-2 (see \url{https://en.wikipedia.org/wiki/ISO_3166-2})}
}
\description{
Create a Spatial Sampling Grid
}
