% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_col_names.R
\name{match_col_names}
\alias{match_col_names}
\title{Ensure Consistency Among eBird and BBS Datasets}
\usage{
match_col_names(x)
}
\arguments{
\item{x}{a data frame}
}
\description{
..
}
\keyword{=}
\keyword{e.g.,}
\keyword{internal}
\keyword{x}
