% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_bbs.r
\name{munge_bbs}
\alias{munge_bbs}
\title{Subset BBS observations data by one or more species of interest.}
\usage{
munge_bbs(
  list = bbs.orig,
  spp = "Double-crested Cormorant",
  zero.fill = TRUE,
  active.only = TRUE,
  keep.stop.level.data = FALSE,
  QualityCurrentID = 1,
  collapse = TRUE,
  countrynums.keep = c(124, 840),
  countrynums.remove = NULL,
  statenums.keep = NULL,
  statenums.remove = NULL
)
}
\arguments{
\item{list}{A list with element "species_list", obtained from running bbsAssistant::get_bbs_data()...}

\item{spp}{A vector of one or more species (using English Common Name) to subset the data by. Capitalization ignored.}

\item{zero.fill}{If TRUE and a single species is provided in 'spp', this function will output list$observations with zero-filled data.}

\item{active.only}{Logical. If TRUE keep only active routes. Discontinued routes will be discarded.}
}
\description{
...
}
