% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_bbs_shapefiles.R
\name{munge_bbs_shapefiles}
\alias{munge_bbs_shapefiles}
\title{Manipulate the BBS route shapefiles}
\usage{
munge_bbs_shapefiles(
  cws.routes.dir,
  usgs.routes.dir,
  cws.layer = "ALL_ROUTES",
  usgs.layer = "US_BBS_Route-Paths-Snapshot_Taken-Feb-2020",
  proj.target = "USGS"
)
}
\arguments{
\item{cws.routes.dir}{Directory for where the CWS (Canadian BBS) shapefiles are stored.}

\item{usgs.routes.dir}{Directory for where the USGS (USA BBS) shapefiles are stored.}

\item{cws.layer}{Name of the layer to import. Defaults to "ALL_ROUTES"}

\item{usgs.layer}{Name of the layer to import.}

\item{proj.target}{One of c("USGS", "CWS"). The remaining datafile will be transformed to match the projection of proj.target.}
}
\description{
This function is meant to intake two shapefiles, one per cws and usgs munges them such that they conform to the USGS data release for BBS observations and route metadata.
}
