% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-munge_col_names.R
\name{munge_col_names}
\alias{munge_col_names}
\title{Munge Column Names for Bird Observations Data}
\usage{
munge_col_names(data)
}
\arguments{
\item{data}{flat data object with column names}
}
\description{
Munge Column Names for Bird Observations Data
}
