% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_date_time.R, R/utils.R
\name{munge_date_time}
\alias{munge_date_time}
\title{Munge Dates and Times of the eBird and BBS Datasets}
\usage{
munge_date_time(dat, base.date, min.yday = 0, max.yday = 365, sunlight = FALSE)

munge_date_time(dat, base.date, min.yday = 0, max.yday = 365, sunlight = FALSE)
}
\arguments{
\item{dat}{data set with at least the column \code{date} and of class "date"}

\item{base.date}{character or date string (YYYY-MM-DD) to use as the origin date for calculating Julian date.}

\item{min.yday}{minimum day of the year to include in resulting dataset}

\item{max.yday}{maximum day of the year to include in resulting dataset}

\item{sunlight}{logical If TRUE will calculate all sunlight, moonlight, rise and set times. This is computationally demanding for the eBird data, so do not set to TRUE unless needed.}
}
\description{
Munge Dates and Times of the eBird and BBS Datasets

Munge Dates and Times of the eBird and BBS Datasets
}
