% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_ebird.R
\name{munge_ebird}
\alias{munge_ebird}
\title{Munge eBird Data Using Auk}
\usage{
munge_ebird(
  dir.ebird.in,
  dir.ebird.out,
  interest.species = NULL,
  interest.spatial = NULL
)
}
\arguments{
\item{dir.ebird.out}{Where to store the munged versions of the data.}

\item{dir.ebd.in}{Directory/directories for where the \strong{unpacked} EDB data file(s) are stored.}
}
\description{
Wrapper for handling eBird data using package Auk.
}
