% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_big_fkn_fun.R
\name{my_big_fucking_function}
\alias{my_big_fucking_function}
\title{My big fkn fun}
\usage{
my_big_fucking_function(
  dir.orig.data,
  dir.proj = NULL,
  species = c("Double-crested Cormorant", "Nannopterum auritum"),
  species.abbr = c("DCCO", "doccor"),
  countries = c("US", "CA"),
  states = NULL,
  year.range = 2008:2019,
  base.julian.date = lubridate::ymd(paste0(min(year.range), c("-01-01"))),
  crs.target = 4326,
  get.sunlight = FALSE,
  grid.size = 1,
  hexagonal = TRUE,
  overwrite.grid = FALSE,
  usgs.layer = "US_BBS_Route-Paths-Snapshot_Taken-Feb-2020",
  cws.layer = "ALL_ROUTES",
  overwrite.bbs = FALSE,
  overwrite.ebird = FALSE,
  max.C.ebird = 100,
  remove.bbs.obs = TRUE,
  max.effort.km = 5,
  max.num.observers = 10,
  max.effort.mins = 180,
  complete.checklists.only = TRUE,
  ebird.protocol = c("Traveling", "Stationary"),
  min.yday = 91,
  max.yday = 245,
  mmyyyy = "sep-2021",
  jagam.args = list(bs = "ds", k = 20, family = "poisson", sp.prior = "log.uniform",
    diagonalize = TRUE),
  scale.vars = TRUE
)
}
\arguments{
\item{dir.orig.data}{tba}

\item{dir.proj}{tba}

\item{species}{tba}

\item{species.abbr}{tba}

\item{countries}{tba}

\item{states}{tba}

\item{year.range}{tba}

\item{base.julian.date}{tba}

\item{crs.target}{tba}

\item{get.sunlight}{tba}

\item{grid.size}{tba}

\item{hexagonal}{tba}

\item{overwrite.grid}{tba}

\item{usgs.layer}{tba}

\item{cws.layer}{tba}

\item{overwrite.bbs}{tba}

\item{max.C.ebird}{tba}

\item{remove.bbs.obs}{tba}

\item{max.effort.km}{tba}

\item{max.num.observers}{tba}

\item{max.effort.mins}{tba}

\item{complete.checklists.only}{tba}

\item{ebird.protocol}{tba}

\item{min.yday}{tba}

\item{max.yday}{tba}

\item{mmyyyy}{tba}

\item{jagam.args}{tba}

\item{scale.vars}{tba}
}
\description{
tba
}
