% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_ebird_events.R
\name{partition_ebird_events}
\alias{partition_ebird_events}
\title{Partition the eBird Sampling Events Data File by States/Provinces}
\usage{
partition_ebird_events(
  dir.ebird.in,
  mmyyyy,
  overwrite = FALSE,
  countries = c("US", "CA", "MX"),
  ncores = NULL
)
}
\arguments{
\item{dir.ebird.in}{directory path to where ebird sampling events is/are saved}

\item{mmyyyy}{month and year associated with the sampling events file.}

\item{countries}{list of 2-letter country codes. Only sampling events from these countries will be partitioned into indiviudal sampling event files on local disk. Specify countries=NULL to partition all countries (this will take a while)}

\item{cols.remove}{which columns will be removed upon import. For internal use primarily. Changing my disrupt the workflow as this feature has not been tested downstream.}
}
\description{
Relies heavily on package 'data.table' to import, filter, and write the sampling events data for each country of interest.
}
