% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pluck_multiple.R
\name{pluck_multiple}
\alias{pluck_multiple}
\title{Pluck Multiple Elements from List Object}
\source{
https://stackoverflow.com/questions/46983716/does-a-multi-value-purrrpluck-exist
}
\usage{
pluck_multiple(x, keep = NULL, remove = NULL)
}
\arguments{
\item{x}{list object}

\item{keep}{character or integer vector of one or more list element names or numbers to include in list output}

\item{remove}{character or integer vector of one or more list element names or numbers to remove from list output}
}
\description{
Pluck Multiple Elements from List Object
}
