% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj.shorthand}
\alias{proj.shorthand}
\title{Auto generate a subdirectory name based on project parameters}
\usage{
\method{proj}{shorthand}(species, states, grid.size, year.range)
}
\arguments{
\item{species}{list of species names. will use the longest value as the name}

\item{states}{list of state/province names. will use the shortest values in the name}

\item{year.range}{Vector of years. will take the min and max value}
}
\description{
Auto generate a subdirectory name based on project parameters
}
