% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_in_jags.R
\name{run_in_jags}
\alias{run_in_jags}
\title{Run BUGS Model in JAGS}
\usage{
run_in_jags(
  bugs.data,
  model,
  inits,
  monitor,
  savedir = "/",
  seed = sample(1:111111, size = 1),
  parallel = TRUE,
  verbose = TRUE,
  mod.name = "myJAGSModel",
  overwrite = FALSE,
  traceplots = TRUE,
  mcmc.specs = set_mcmc_specs()
)
}
\arguments{
\item{bugs.data}{list List of elements comprising constants and data for use in JAGS}

\item{model}{Filepath to file containing JAGS model}

\item{inits}{list List of initial values.}

\item{monitor}{optional Character vector of parameters to monitor.}

\item{savedir}{If not specified, will save the samples resulting from jagsUI::jags() to current working directory}

\item{seed}{optional If specified will set a seed for random number generation.}

\item{parallel}{logical If TRUE will run in parallel}

\item{verbose}{logical Argument used in jagsUI::jags(). If TRUE will generate more messages/information during sampling phase. Note: If parallel=TRUE, messages are suppressed given behavior of parallel compute.}

\item{mod.name}{optional Used to save model output to file. Defaults to 'myJAGSModel'}

\item{overwrite}{logical If TRUE and .RDS file already exists, will prompt user with a menu to confirm model re-run. Specifying overwrite=TRUE will avoid that prompt.}

\item{traceplots}{logical If TRUE will save traceplots to file}

\item{mcmc.specs}{list List of specs for use in MCMC specification (e.g., n.chains, n.adapt, n.cores). If any necessary values are missing, will use set_mcmc_specs to grab a default value.}
}
\description{
Run BUGS Model in JAGS
}
