% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_in_nimble.r
\name{run_in_nimble}
\alias{run_in_nimble}
\title{Run BUGS Model in nimble}
\usage{
run_in_nimble(
  myData,
  myModel,
  myInits = NULL,
  traceplots = TRUE,
  constants = NULL,
  monitors = NULL,
  savedir = "outputs",
  seed = sample(1:111111, size = 1),
  parallel = FALSE,
  verbose = TRUE,
  mod.name = "mynimbleModel",
  mcmc.specs = set_mcmc_specs(dev.mode = FALSE),
  ...
)
}
\arguments{
\item{myData}{list List of elements comprising data (and constants if not provided separately) for use in nimble}

\item{myModel}{nimbleCode object or filepath to a BUGS model.}

\item{myInits}{list List of initial values.}

\item{traceplots}{logical If TRUE will attempt to print default traceplots for model output}

\item{constants}{optional Can be provided separately from myData. If both are provided, function will combine prior to inclusion in readNimbleModel}

\item{monitors}{optional Character vector of parameters to monitor.}

\item{savedir}{If not specified, will save the samples resulting from nimbleUI::nimble() to current working directory}

\item{seed}{optional If specified will set a seed for random number generation.}

\item{parallel}{logical If TRUE will run in parallel}

\item{verbose}{logical Argument used in nimble::compileNimble(showCompilerOutput). If TRUE will generate more messages/information during sampling phase. Note: If parallel=TRUE, messages are suppressed given behavior of parallel compute.}

\item{mod.name}{optional Used to save model output to file. Defaults to 'mynimbleModel'}

\item{mcmc.specs}{list List of specs for use in MCMC specification (e.g., n.chains, n.adapt, n.cores). If any necessary values are missing, will use set_mcmc_specs to grab a default value.}

\item{...}{Additional arguments}
}
\description{
Run BUGS Model in nimble
}
