% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC_allcode.R
\name{run_MCMC_allcode}
\alias{run_MCMC_allcode}
\title{Run MCMC in or Out of Parallel in nimble}
\usage{
run_MCMC_allcode(
  data,
  constants = NULL,
  model,
  inits,
  mcmc.specs,
  verbose = FALSE,
  seed = sample(1:1e+06, 1),
  monitors = NULL,
  ...
)
}
\arguments{
\item{data}{data as a list}

\item{constants}{constants as list}

\item{model}{filepath or nimbleModel object}

\item{inits}{list of initial values}

\item{mcmc.specs}{MCMC specifications as a list; created using \code{set_mcmc_specs}}

\item{verbose}{logical if TRUE and code is not running in parallel will print all output available from Nimble commands.}

\item{seed}{for internal use}

\item{monitors}{list of parameters to monitor}

\item{...}{Additional arguments}
}
\description{
Run MCMC in or Out of Parallel in nimble
}
