% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run JAGS Model}
\usage{
run_model(
  jdat,
  mod.fn,
  mod.name = "mymodel",
  model.out.dir = NULL,
  inits = NULL,
  mcmc.specs = set_mcmc_specs(),
  use.dclone = TRUE,
  params.monitor = NULL,
  make.plot = FALSE,
  plot.dir = NULL,
  view.plot = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{jdat}{Data input to JAGS}

\item{mod.fn}{Local file path for JAGS model file.}

\item{mod.name}{Model name. Used to name output files.}

\item{model.out.dir}{Directory path where to save output files (e.g., mod.name.RDS)}

\item{inits}{Initial values for a single chain. If initial values are not provided for all chains, this function will repeat the intiial values provided .}

\item{mcmc.specs}{Specifications for MCMC chains. Should be created using bbsebird::set_mcmc_species. If  not provided will default to default values of bbsebird::set_mcmc_specs}

\item{use.dclone}{logical If TRUE will use dclone::jags.parfit to fit model in JAGS. See package dclone for more details.}

\item{params.monitor}{Vector List of parameters to monitor.}

\item{plot.dir}{Directory path where to save output figures.}

\item{view.plot}{logical IF TRUE will open the plot saved to file}

\item{overwrite}{logical If FALSE and mod.name.RDS already exists in model.out.dir, will prompt user to confirm whether they wish to import existing model outputs or overwrite.}

\item{plot}{logical If TRUE will produce default traceplots of model output}
}
\description{
Run JAGS Model
}
