% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nimble_model.R
\name{run_nimble_model}
\alias{run_nimble_model}
\title{Run Nimble Model in Parallel
Function will run a Nimble model for 'nc' indepdendent chains}
\usage{
run_nimble_model(
  code,
  data,
  constants = NULL,
  inits,
  monitors = NULL,
  ncores = NULL,
  ni = 10000,
  nb = NULL,
  nt = 1,
  nc = 1,
  aI = 200,
  ntries = 5,
  calculate = FALSE,
  block.name = "alpha+b",
  block.samp.type = "RW_block",
  parallel = TRUE
)
}
\arguments{
\item{code}{nimble model code}

\item{data}{data as a list}

\item{constants}{constants as list}

\item{inits}{list of initial values}

\item{monitors}{optional Character vector of parameters to monitor.}

\item{ncores}{maximum number of cores to employ. Actual number used is the minimum of nc and ncores}

\item{ni}{number iterations to run}

\item{nb}{number of burn-in iterations to discard (I think it's PRE-THINNING burnin discard...)}

\item{nt}{thinning rate (every Nth iteration will be saved)}

\item{nc}{number of chains to run (in parallel)}

\item{aI}{adapt interval, used in nimble::addSamplers}

\item{ntries}{optional If using parameter block sampler, specify the maximum number of tries}

\item{calculate}{logical if TRUE will calculate the model logprob. Used as argument 'calculate' in function nimble::nimbleModel()}

\item{block.name}{optional one of c("alpha+b", "all"). If "alpha+b" will block each alpha and b across all T. If "all" will block all alpha and b for each Ts.}

\item{block.samp.type}{optional one of c("AF_slice", "RW_block").}

\item{parallel}{logical if TRUE will run chains in parallel (using foreach).}
}
\description{
Run Nimble Model in Parallel
Function will run a Nimble model for 'nc' indepdendent chains
}
