% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_par_nimble.R
\name{run_par_nimble}
\alias{run_par_nimble}
\title{Run Nimble Model in Parallel
Function will run a Nimble model for 'nc' indepdendent chains}
\usage{
run_par_nimble(
  code,
  data,
  constants = NULL,
  inits,
  monitors = NULL,
  ncores = NULL,
  ni = 10000,
  nb = NULL,
  nt = 1,
  nc = 1,
  aI = 100,
  ntries = 5,
  block.name = NULL,
  block.samp.type = "AF_slice",
  mod.name = "mymodel",
  seed = 123
)
}
\arguments{
\item{code}{nimble model code}

\item{data}{data as a list}

\item{constants}{constants as list}

\item{inits}{list of initial values}

\item{monitors}{optional Character vector of parameters to monitor.}

\item{ncores}{maximum number of cores to employ. Actual number used is the minimum of nc and ncores}

\item{ni}{number iterations to run}

\item{nb}{number of burn-in iterations to discard}

\item{nt}{thinning rate (every Nth iteration will be saved)}

\item{nc}{number of chains to run (in parallel)}

\item{aI}{optional If using parameter block sampler, specify the adaptive interval}

\item{ntries}{optional If using parameter block sampler, specify the maximum number of tries}

\item{block.name}{optional one of c("alpha+b", "all"). If "alpha+b" will block each alpha and b across all T. If "all" will block all alpha and b for each Ts.}

\item{block.samp.type}{optional one of c("AF_slice", "RW_block").}

\item{mod.name}{model name}

\item{seed}{used to set the seed for nimble within each processor}
}
\description{
Run Nimble Model in Parallel
Function will run a Nimble model for 'nc' indepdendent chains
}
