% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_list_dims.r
\name{save.runtime}
\alias{save.runtime}
\title{Write Model Runtimes to Working Directory}
\usage{
save.runtime(name, time, fn.mod = "mymodel.rds", fn = "model-runtimes.txt")
}
\description{
Write Model Runtimes to Working Directory
}
