% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Computes standard error of a vector
}
\examples{
x <- rnorm(100)
se(x)
}
\keyword{error}
\keyword{standard}
