% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.filenames}
\alias{set.filenames}
\title{Specify Model Output Filenames}
\usage{
set.filenames(
  mod.name,
  dir.proj = getwd(),
  dir.plots = NULL,
  dir.samps = NULL,
  ni = NULL,
  nb = NULL,
  nc = NULL,
  nt = NULL,
  nbfs = NULL,
  dev.mode = FALSE
)
}
\arguments{
\item{mod.name}{a shorthand name for saving outputs.}

\item{dir.proj}{location of where to send exported files. If NULL or not specified will send to current working directory.}

\item{dir.plots}{path to save the MCMC samples as .rds object}

\item{dir.samps}{path to save the MCMC samples as .rds object}

\item{ni}{number of iterations (pre-thinning and pre-burnin)}

\item{nb}{number of iterations to discard}

\item{nc}{number of chains}

\item{nt}{thinning rate for MCMC. Used in filenames}

\item{nbfs}{number of basis functions used in the model.}

\item{dev.mode}{logical If TRUE will append "dev_" preceding all filenames.}
}
\description{
Helper function for creating filenames for exporting MCMC samples and plots thereof.
}
