% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_args.R
\name{set_args}
\alias{set_args}
\title{Specify Arguments and Directories}
\usage{
set_args(
  dir.orig.data,
  dir.proj,
  species,
  species.abbr,
  states = NULL,
  countries = NULL,
  year.range = 2008:2019,
  crs.target = 4326,
  get.sunlight = FALSE,
  base.julian.date = (paste0(min(year.range), c("-01-01"))),
  grid.size = 1,
  hexagonal = TRUE,
  overwrite.grid = FALSE,
  usgs.layer = "US_BBS_Route-Paths-Snapshot_Taken-Feb-2020",
  cws.layer = "ALL_ROUTES",
  overwrite.bbs = FALSE,
  overwrite.ebird = FALSE,
  max.C.ebird = 100,
  remove.bbs.obs = TRUE,
  max.effort.km = 5,
  max.num.observers = 10,
  max.effort.mins = 180,
  complete.checklists.only = TRUE,
  ebird.protocol = c("Traveling", "Stationary"),
  min.yday = 91,
  max.yday = 245,
  mmyyyy = "nov-2021",
  overwrite.jdat = FALSE,
  jagam.args = list(bs = "ds", k = 20, family = "poisson", sp.prior = "log.uniform",
    diagonalize = TRUE),
  scale.vars = TRUE
)
}
\description{
Function to specify arugments and create/point to directories based on those arguments.
}
