% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-and-inits.R
\name{set_mcmc_specs}
\alias{set_mcmc_specs}
\title{Set Specifications for MCMC}
\usage{
set_mcmc_specs(
  na = 1000,
  nb = 5000,
  nc = 3,
  ni = 20000,
  nt = 10,
  in.parallel = FALSE,
  ncores = parallel::detectCores() - 1,
  dev.mode = 2
)
}
\arguments{
\item{na}{number of adaptation iterations}

\item{nb}{number of burn-in iterations per chain}

\item{nc}{number of chains}

\item{ni}{number of iterations per chain}

\item{nt}{thinning rate (every nth iteration removed)}

\item{in.parallel}{logical if TRUE will set number chains ('nc') value to 1}

\item{ncores}{number of cores to run model in parallel}

\item{dev.mode}{integer 1:3. 1=large-ish run; 2=moderate run; 3=quick, short run.}
}
\description{
Set Specifications for MCMC
}
