% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bbs.R
\name{sim_bbs}
\alias{sim_bbs}
\title{Simulate BBS Data for Use in JAGS}
\usage{
sim_bbs(
  nyear = 20,
  nsite = 10,
  ngrid = 10,
  zip = TRUE,
  maxN = 33,
  probMinMultG = 0.75,
  probfyear = 0.15,
  propna = 0.25,
  maxgridperroute = 5
)
}
\arguments{
\item{nyear}{number of years}

\item{nsite}{number of routes}

\item{ngrid}{number of grid cells}

\item{zip}{TRUE will generate N using a zero-inflated poisson distribution, FALSE non-zip}

\item{maxN}{max number in the true state}

\item{probMinMultG}{probability of a route falling into multiple grid cells. probMinMultG should decrease as grid cell size increases (i.e., less chance of a rotue falling into multiple grid cells as spatial coverage of a cell increases)}

\item{propna}{proportion of all routes by year combinations to have no data (NA). this is a random process so will not always equal propna (uses rbinom to generate NAs)}

\item{maxgridperroute}{max number of grids that a single route can span. use whatever makes most sense. if probMinMultG is very low, then use 1}

\item{propfyer}{proportion of routes by year combinations that comprise an observer's first year}
}
\description{
Creates a list of BBS data for use in JAGS. Data is simulated using Poisson or zero-inflated Poisson distributions. Output includes both N and C.
}
