% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate BBS Data for Use in JAGS}
\usage{
sim_data(
  nyear = 20,
  nroutes = 100,
  nchecklistsmaxperyear = 100,
  ngrid = 10,
  zip = TRUE,
  maxNb = 33,
  maxNe = 25,
  probMinMultG = 0.75,
  probChecklistInGrid = 0.85,
  probfyear = 0.15,
  propna = 0.25,
  maxgridperroute = 5,
  show.plot = TRUE
)
}
\arguments{
\item{nyear}{number of years}

\item{nroutes}{number of BBS routes sampled across all years}

\item{ngrid}{number of grid cells}

\item{zip}{TRUE will generate N using a zero-inflated poisson distribution, FALSE non-zip}

\item{maxNb}{max number in the true state for BBS}

\item{maxNe}{max number in the true state for eBird}

\item{probMinMultG}{probability of a route falling into multiple grid cells. probMinMultG should decrease as grid cell size increases (i.e., less chance of a rotue falling into multiple grid cells as spatial coverage of a cell increases)}

\item{probChecklistInGrid}{probability of a gridcell in a given year having data(checklistid). Setting this below 1.0 allows for some cells to be without checklists each year}

\item{propna}{proportion of all routes by year combinations to have no data (NA). this is a random process so will not always equal propna (uses rbinom to generate NAs)}

\item{maxgridperroute}{max number of grids that a single route can span. use whatever makes most sense. if probMinMultG is very low, then use 1}

\item{show.plot}{if TRUE will print a density plot of N and C to device}

\item{propfyer}{proportion of routes by year combinations that comprise an observer's first year}

\item{nchecklistmaxperyear}{maximum number of ebird checklists to randomly impute into grid cells per year}
}
\description{
Creates a list of BBS data for use in JAGS. Data is simulated using Poisson or zero-inflated Poisson distributions. Output includes both N and C.
}
