% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_rmd_chunks.R
\name{source_rmd_chunks}
\alias{source_rmd_chunks}
\title{Source Select Chunks Inside Another R Markdown File}
\usage{
source_rmd_chunks(file, chunk_labels, skip_plots = TRUE, output_temp = FALSE)
}
\arguments{
\item{file}{Rmarkdown filename to hit up}

\item{chunk_labels}{chunk labels to run}
}
\description{
This function was borrowed from https://gist.githubusercontent.com/brshallo/e963b9dca5e4e1ab12ec6348b135362e/raw/71ad2b7288dc6e4f277f53f00e650d98a8e54aff/source-rmd-chunks.r
}
