% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trySave.R
\name{trySave}
\alias{trySave}
\title{Try to Save Results or plot(results) to File}
\usage{
trySave(
  x,
  name,
  savedir = "/",
  mcmc.specs = NULL,
  traceplots = FALSE,
  monitors = "all"
)
}
\arguments{
\item{x}{object to save to file as .RDS or as .PDF}

\item{name}{object name (e.g., plots-nimble or jags_samples)}

\item{savedir}{directory within which object will attempt to save}

\item{mcmc.specs}{if provided, the MCMC specs will be used to save output files}

\item{traceplots}{logical If TRUE will save plots to PDF}

\item{monitors}{parameters monitored and that user wishes to plot}
}
\description{
Try to Save Results or plot(results) to File
}
