% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-model-output.R
\name{trySaveModel}
\alias{trySaveModel}
\title{Ensure JAGS Model Outputs are Saved
Sometimes when working in OneDrive/Dropbox or on company networks, you may ahve problems saving to file. Running this instead of directly saving will ensure model output (.rds) is saved to an alternate location. This may be handy for running models in parallel/loops.
The function returns the output filename. Will save to a directory within the user's Desktop (windows, mac) if alt.save.dir is unspecified}
\usage{
trySaveModel(
  obj,
  obj.name = "mymodel",
  save.dir,
  alt.save.dir = paste0(file.path(Sys.getenv("USERPROFILE"), "Desktop"),
    "/JAGS.outputs/"),
  alt.save.subdir = NULL
)
}
\description{
Ensure JAGS Model Outputs are Saved
Sometimes when working in OneDrive/Dropbox or on company networks, you may ahve problems saving to file. Running this instead of directly saving will ensure model output (.rds) is saved to an alternate location. This may be handy for running models in parallel/loops.
The function returns the output filename. Will save to a directory within the user's Desktop (windows, mac) if alt.save.dir is unspecified
}
