% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_helpers.R
\name{unpack_ebird}
\alias{unpack_ebird}
\title{unpack_ebird}
\usage{
unpack_ebird(
  dir = "data-raw/ebird-data",
  overwrite = FALSE,
  spp.ind = "doccor"
)
}
\arguments{
\item{dir}{Directory where the packed ebird data files are stored and into which files will unpack.}

\item{overwrite}{Logical. Default=FALSE will not overwrite existing files in dir.}

\item{spp.ind}{A species indicator (using eBird terminology), used to identify which .zip files should be unpacked. Defaults to 'doccor' (Double-crested Cormorant).}
}
\description{
Unpacks the eBird files (EBD and SamplingEvent data). Saves all unpacked data to a subdirectory (see: parameter dir)
}
