% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerofill-ebird.R
\name{zerofill_ebird}
\alias{zerofill_ebird}
\title{Create a Zero-filled Data Object for eBird Observations}
\usage{
zerofill_ebird(
  myList,
  keep.orig = FALSE,
  cols.remove = c("SUBSPECIES COMMON NAME", "TAXONOMIC ORDER", "LAST EDITED DATE",
    "CATEGORY", "APPROVED", "REVIEWED", "SPECIES COMMENTS", "...48", "HAS MEDIA",
    "REASON", "TRIP COMMENTS")
)
}
\arguments{
\item{myList}{A list containing two named data frames, c("observations", "sampling"). This object is the result of \code{filter_ebird_data()}.}

\item{keep.orig}{Logical. If FALSE will delete the original object, myList, from memory.}

\item{cols.remove}{A vector of column names to be excluded from the output file.}

\item{cols.to.lowercase}{Logical. If TRUE will export a data frame where all colnames are in lowercase. Capitalization does not matter.}
}
\description{
Creates a zero-filled data object comprising the eBird observations and sampling events data supplied.
}
