% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsebird-package.R
\docType{package}
\name{bbsebird}
\alias{bbsebird}
\alias{bbsebird-package}
\title{bbsebird: Unifying BBS and eBird Data}
\description{
Imports, munges and aligns eBird and BBS data to a spatial sampling grid. This package also outputs data in various forms for use in JAGS.
}
\author{
\strong{Maintainer}: Jessica Burnett \email{jburnett@usgs.gov} (\href{https://orcid.org/0000-0002-0896-5099}{ORCID})

}
\keyword{internal}
