% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data.r
\name{bundle_data}
\alias{bundle_data}
\title{Bundle Data for Use in JAGS (LONG format)}
\usage{
bundle_data(bbs_spatial, ebird_spatial, grid)
}
\arguments{
\item{bbs_spatial}{BBS data}

\item{ebird_spatial}{eBird data}

\item{grid}{spatial sampling grid/study area}
}
\description{
An updated version of bundle_data, where site by year elements are provided in long format
}
