% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data.r
\name{bundle_data}
\alias{bundle_data}
\title{Bundle Data for Use in JAGS (LONG format)}
\usage{
bundle_data(
  bbs,
  ebird,
  grid,
  scale.covs = TRUE,
  X = "cell.lon.centroid",
  Y = "cell.lat.centroid",
  cell.id = "gridcellid",
  year.id = "year",
  site.id = c("checklist_id", "rteno"),
  obs.id = c("obsn", "observer_id"),
  cell.covs = c("area"),
  site.covs = c("wind", "noise", "cars", "observ", "minute", "assistant",
    "obsfirstyearbbs", "obsfirstyearroute", "duration_minutes", "effort_distance_km",
    "effort_area_ha", "number_observers"),
  K = NULL
)
}
\arguments{
\item{bbs}{BBS data}

\item{ebird}{eBird data}

\item{grid}{spatial sampling grid/study area}

\item{scale.covs}{logical if TRUE will automatically scale the covariates}

\item{X}{variable name associated with the x-coordinate (e.g., long, longitude, Easting, X) across 'grid', 'bbs', and 'ebird'}

\item{Y}{variable name associated with the x-coordinate (e.g., latitude, Northing, Y) across 'grid', 'bbs', and 'ebird'}

\item{cell.id}{column name(s) of the grid cell identifier}

\item{year.id}{column name of the temporal identifier}

\item{site.id}{column name(s) of the site  identifier (e.g., BBS route, eBird checklists)}

\item{obs.id}{column name(s) of the observer identifier}

\item{cell.covs}{column name(s) of the grid-level covariates}

\item{site.covs}{column name(s) of the site-level covariates}

\item{K}{the maximum number of basis functions that JAGAM will produce. Used for code development purposes, mostly. Do not change unless you know what you're doing.}
}
\description{
An updated version of bundle_data, where site by year elements are provided in long format
}
