% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_data.R
\name{bundle_data}
\alias{bundle_data}
\title{Bundle Data for Use in JAGS (LONG format)}
\usage{
bundle_data(
  bbs,
  ebird,
  grid,
  scale.covs = TRUE,
  fill.cov.nas = NULL,
  bf.method = "cubic2D",
  use.ebird.in.ENgrid = TRUE,
  ENgrid.arg = "max",
  X = "cell.lon.centroid",
  Y = "cell.lat.centroid",
  cell.id = "gridcellid",
  year.id = "year",
  site.id = c("checklist_id", "rteno"),
  obs.id = c("obsn", "observer_id"),
  cell.covs = c("area"),
  site.covs = c("wind", "noise", "cars", "observ", "minute", "assistant",
    "obsfirstyearbbs", "obsfirstyearroute", "duration_minutes", "effort_distance_km",
    "effort_area_ha", "number_observers"),
  K = NULL,
  dev.mode = FALSE
)
}
\arguments{
\item{bbs}{BBS data}

\item{ebird}{eBird data}

\item{grid}{spatial sampling grid/study area}

\item{scale.covs}{logical if TRUE will automatically scale the numeric/integer covariates.}

\item{fill.cov.nas}{value (e.g., 666, 99) or function ("max", "min", "mode", "mean") with which to fill NA values on covariate data. Defaults to NULL, which does nothing.}

\item{bf.method}{method for developing basis functions. Defaults to creating duchon splines using mgcv::jagam().}

\item{use.ebird.in.ENgrid}{logical if TRUE will use data across both eBird and BBS observations to create basis functions.}

\item{ENgrid.arg}{if "max" will use the maximum value of observed birds at each grid cell. Alternatives include "min", "mean".}

\item{X}{variable name associated with the x-coordinate (e.g., long, longitude, Easting, X) across 'grid', 'bbs', and 'ebird'}

\item{Y}{variable name associated with the x-coordinate (e.g., latitude, Northing, Y) across 'grid', 'bbs', and 'ebird'}

\item{cell.id}{column name(s) of the grid cell identifier}

\item{year.id}{column name of the temporal identifier}

\item{site.id}{column name(s) of the site  identifier (e.g., BBS route, eBird checklists)}

\item{obs.id}{column name(s) of the observer identifier}

\item{cell.covs}{column name(s) of the grid-level covariates}

\item{site.covs}{column name(s) of the site-level covariates}

\item{K}{the maximum number of basis functions that JAGAM will produce. Used for code development purposes, mostly. Do not change unless you know what you're doing.}

\item{dev.mode}{logical if TRUE will return a reduced data set to use in development/debugging purposes. This method reduces the number of time units to 2, the maximum number of grid cells to 10, and 2 sites from each data source}
}
\description{
An updated version of bundle_data, where site by year elements are provided in long format
}
