% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_params}
\alias{eval_params}
\title{Spruce up R Markdown YAML parameter values that are lists into clean vectors. Called for convenience inside .RMD files.}
\usage{
eval_params(x = params)
}
\arguments{
\item{x}{The parameter object from YAML \code{params}}
}
\description{
Spruce up R Markdown YAML parameter values that are lists into clean vectors. Called for convenience inside .RMD files.
}
\keyword{internal}
