% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ebird_data.R
\name{filter_ebird_data}
\alias{filter_ebird_data}
\title{Create and Write or Load In the Filtered eBird Data}
\usage{
filter_ebird_data(
  fns.ebird,
  dir.ebird.out,
  countries = NULL,
  states = NULL,
  complete.only = TRUE,
  protocol = c("Traveling", "Stationary"),
  species = "Double-crested Cormorant",
  overwrite = FALSE,
  remove.bbs.obs = TRUE,
  years = NULL,
  max.effort.km = NULL,
  max.effort.mins = NULL,
  max.num.observers = 10,
  method = "data.table",
  cols.keep = c("all_species_reported", "country", "common_name", "country_code",
    "county", "county_code", "duration_minutes", "effort_area_ha", "effort_distance_km",
    "group_identifier", "latitude", "longitude", "number_observers", "observation_date",
    "observation_count", "observer_id", "protocol_code", "protocol_type",
    "sampling_event_identifier", "state", "scientific_name", "state_code",
    "time_observations_started"),
  f_obs_out = paste0(dir.ebird.out, "ebird_obs_filtered.txt"),
  f_samp_out = paste0(dir.ebird.out, "ebird_samp_filtered.txt")
)
}
\arguments{
\item{fns.ebird}{File paths for the EBD and SamplingEvents data to import. Character vector of filenames for original files.}

\item{dir.ebird.out}{Location of where to save and/or find the filtered/subsetted eBird data.}

\item{overwrite}{logical If true will overwrite existing filtered data objects in directory \code{dir.ebird.out}}

\item{f_obs_out}{Filename for saving the filtered eBird observations data}

\item{f_samp_out}{Filename for saving the filtered eBird sampling data}
}
\description{
Filter the eBird data and sampling events using R package AUK.
}
