% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_structure.R
\name{get_data_structure}
\alias{get_data_structure}
\title{Get lista Structure of JAGS lista Inputs}
\usage{
get_data_structure(list, filename = NULL)
}
\arguments{
\item{list}{A list of named objects.}

\item{filename}{If specified saves the resulting metadata table to file as .csv}
}
\description{
Get lista Structure of JAGS lista Inputs
}
