% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_ebird_files.R
\name{id_ebird_files}
\alias{id_ebird_files}
\title{...}
\usage{
id_ebird_files(
  dir.ebird.in,
  mmyyyy = "oct-2021",
  regions = NULL,
  species = "doccor",
  country.code.identifier = "iso2c",
  sampling.events = TRUE,
  get.full.data = FALSE
)
}
\arguments{
\item{dir.ebird.in}{Directory for where the ebird data are stored.}

\item{mmyyyy}{The month (mm) and year (yyyy) of the ebird dataset you wish to use. Suggested to use most recent on file.}

\item{regions}{Character vector comprising state/province or country abbreviations. If not specified, will identify the files associated with the country-level data.}

\item{species}{Character vector comprising species identifiers.}

\item{country.code.identifier}{Identifier used by eBird to define countries. Should not be changed unless eBird changes its practice of using the iso2c}

\item{sampling.events}{Logical.}

\item{get.full.data}{Logical if TRUE will retrieve the filenames for the ENTIRE database, thereby ignoring the species and regions.}
}
\description{
...
}
