% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jdat.contents}
\alias{jdat.contents}
\title{Descriptions of JDAT Contents}
\format{
A data frame
\describe{
\item{level.1.name}{name of the top level list element}
\item{level.2.name}{name of the second-level list element}
\item{level.1.description}{description of the top level list element}
\item{level.2.description}{description of the second-level list element}
\item{class}{expected class of object}
\item{notes}{supplementary information}
\item{length}{if lowest level object is a vector, length of object}
\item{rows}{if object is array, number of rows}
\item{c}{if object is array, number of cols}
\item{s}{if object is 3-D array, number of slices}
}
}
\usage{
jdat.contents
}
\description{
A simple lookup table for
}
\keyword{datasets}
