% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bbs_spatial.R
\name{make_bbs_spatial}
\alias{make_bbs_spatial}
\title{Manipulate the BBS route shapefiles}
\usage{
make_bbs_spatial(
  bbs.obs,
  cws.routes.dir,
  usgs.routes.dir,
  routes.keep = NULL,
  cws.layer = "ALL_ROUTES",
  usgs.layer = "US_BBS_Route-Paths-Snapshot_Taken-Feb-2020",
  crs.target = 4326,
  grid = NULL,
  overwrite = TRUE,
  print.plots = TRUE,
  keep.empty.cells = TRUE,
  plot.dir = NULL
)
}
\arguments{
\item{cws.routes.dir}{Directory for where the CWS (Canadian BBS) shapefiles are stored.}

\item{usgs.routes.dir}{Directory for where the USGS (USA BBS) shapefiles are stored.}

\item{cws.layer}{Name of the layer to import. Defaults to "ALL_ROUTES"}

\item{usgs.layer}{Name of the layer to import.}
}
\description{
This function is meant to intake two shapefiles, one per cws and usgs munges them such that they conform to the USGS data release for BBS observations and route metadata.
}
