% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bundle.R
\name{make_bundle}
\alias{make_bundle}
\title{Bundle Data for Use in JAGS}
\usage{
make_bundle(
  bbs,
  ebird,
  grid,
  drop.na.cov.obs = TRUE,
  mins.to.hours = TRUE,
  scale.covs = TRUE,
  fill.cov.nas = NA,
  bf.method = "cubic2D",
  use.ebird.in.ENgrid = TRUE,
  ENgrid.arg = "max",
  X = "cell.lon.centroid",
  Y = "cell.lat.centroid",
  cell.id = "gridcellid",
  year.id = "year",
  site.id = c("checklist_id", "rteno"),
  obs.id = c("obsn", "observer_id"),
  cell.covs = c("area"),
  site.covs = c("wind", "noise", "cars", "minute", "assistant", "obsfirstyearbbs",
    "obsfirstyearonbbs", "obsfirstyearroute", "obsfirstyearonroute",
    "time_observations_started_hsm", "time_observations_started", "duration_minutes",
    "protocol_type", "number_observers"),
  K = NULL,
  dev.mode = FALSE,
  dir.jagam = "models/"
)
}
\arguments{
\item{bbs}{BBS data table}

\item{ebird}{eBird data table}

\item{grid}{spatial sampling grid/study area table}

\item{drop.na.cov.obs}{logical if TRUE will remove ALL data where any specified covariate does not exist. If TRUE, suggest examining the data and covariates prior to specifying site.covs and grid.covs.}

\item{mins.to.hours}{logical if TRUE will convert covariates on the minute scale to the hour scale.}

\item{scale.covs}{logical if TRUE will automatically scale the numeric/integer covariates.}

\item{fill.cov.nas}{value with with to fill missing covariate values. User can specify value as FALSE if no fill is requested.}

\item{bf.method}{default value "cubic2d". Alternatives include cubic splines using one of c("mgcv", "jagam").}

\item{use.ebird.in.ENgrid}{logical if TRUE will use data across both eBird and BBS observations to create basis functions.}

\item{ENgrid.arg}{if "max" will use the maximum value of observed birds at each grid cell. Alternatives include "min", "mean".}

\item{X}{variable name associated with the x-coordinate (e.g., long, longitude, Easting, X) across 'grid', 'bbs', and 'ebird'}

\item{Y}{variable name associated with the x-coordinate (e.g., latitude, Northing, Y) across 'grid', 'bbs', and 'ebird'}

\item{cell.id}{column name(s) of the grid cell identifier}

\item{year.id}{column name of the temporal identifier}

\item{site.id}{column name(s) of the site  identifier (e.g., BBS route, eBird checklists)}

\item{obs.id}{column name(s) of the observer identifier}

\item{cell.covs}{column name(s) of the grid-level covariates}

\item{site.covs}{column name(s) of the site-level covariates}

\item{K}{the maximum number of basis functions that JAGAM will produce. Used for code development purposes, mostly. Do not change unless you know what you're doing.}

\item{dev.mode}{logical if TRUE will return a reduced data set to use in development/debugging purposes. This method reduces the number of time units to 2, the maximum number of grid cells to 10, and 2 sites from each data source}

\item{dir.jagam}{directory location of where to save the JAGAM.bugs model file created by mgcv::jagam(). Defaults to ./models/}
}
\description{
Bundle Data for Use in JAGS
}
