% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ebird_spatial.R
\name{make_ebird_spatial}
\alias{make_ebird_spatial}
\title{Create an SF object for eBird Observations}
\usage{
make_ebird_spatial(
  df,
  crs.target = 4326,
  dir.out = NULL,
  grid = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{df}{A data frame with columns c(lat, lon) and optional attributes (columns).}

\item{crs.target}{the integer representing the target CRS.}

\item{dir.out}{path to where the resulting spatial data frame will be saved. If NULL will not save to file.}

\item{grid}{a spatial grid over which the eBird data will be overlaid.}

\item{overwrite}{logical if TRUE will overwrite any existing file named "ebird_spatial.rds" in path dir.out}
}
\description{
Converts eBird (preferably zero-filled data) into a simple features spatial object.
}
