% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gam.R
\name{make_gam}
\alias{make_gam}
\title{Function for Generating Basis Functions}
\usage{
make_gam(
  coords,
  method = "cubic2d",
  nd = NULL,
  K = NULL,
  nruns = 10,
  num.nn = 10,
  ll.to.utm = TRUE,
  max.loop = 40,
  print.plot = TRUE,
  plot.main = NULL,
  scale.coords = TRUE
)
}
\arguments{
\item{coords}{matrix or data frame with easting and northing coordinates, respectively. First column should be easting (e.g., X, long) and second northing.}

\item{method}{one of c("mgcv", "cubic2d", "strebel")}

\item{nd}{number of points to use in cover design. See fields::cover.design.}

\item{K}{max number of basis functions to produce in mgcv::jagam()}

\item{nruns}{The number of random starts to be optimized.See fields::cover.design.}

\item{num.nn}{max number of nearest neighbors. See fields::cover.design.}

\item{ll.to.utm}{logical if TRUE will convert lat long coordinates to UTMs prior to creating distance-based basis functions.}

\item{max.loop}{Maximum number of passes throguh cover.design algorithm. See fields::cover.design.}

\item{print.plot}{logical if TRUE will print spatial representation of site and knot locations to local device.}

\item{plot.main}{optional Character vector for plot title.}

\item{scale.coords}{logical if TRUE will scale and center the XY coordinates using bbsebird::standardize.}
}
\description{
Function for Generating Basis Functions
}
