% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_indexing_df.R
\name{make_indexing_df}
\alias{make_indexing_df}
\title{Make 2-D array for Indexing Purposes}
\usage{
make_indexing_df(X)
}
\arguments{
\item{X}{a data frame with variables site.id, site.ind, grid.id, grid.ind, year.id, year.ind, and "c"}
}
\description{
Make 2-D array for Indexing Purposes
}
\keyword{internal}
