% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jags_list.r
\name{make_jags_list}
\alias{make_jags_list}
\title{Munge and Output a List or List of Lists for Use in JAGS}
\usage{
make_jags_list(
  dat,
  jagam.args,
  max.C.ebird = 100,
  scale.vars = TRUE,
  dir.out,
  fn.out = "jdat"
)
}
\arguments{
\item{dat}{a single data object or a list of multiple objects to munge and collapse into a list of data for use in JAGS}

\item{jagam.args}{List of arguments used in \code{mgcv::jagam()}. Arguments include c(bs, k, family, sp.prior, m, ...)}

\item{max.C.ebird}{if TRUE will drop all checklists where number of observed birds is greater than 100. This is an arbitrary number.}

\item{scale.vars}{Logical If TRUE will scale variables. This needs to be improved/checked.}

\item{dir.out}{Directory within which the output list will be saved  as "jdat"}

\item{fn.out}{Filename of the list object output as a .RDS file}

\item{dir.models}{Location for where to store the GAM portion of the JAGS model}
}
\description{
Munge and Output a List or List of Lists for Use in JAGS
}
