% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{make_list}
\alias{make_list}
\title{Make a List With Named
Take multiple objects from the global environment and place into a single list.}
\usage{
make_list(input)
}
\arguments{
\item{input}{character vector names of data objects in the workspace to be included in the outputted list}
}
\description{
Make a List With Named
Take multiple objects from the global environment and place into a single list.
}
