% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mat.R
\name{make_mat}
\alias{make_mat}
\title{Make 2-D array (Matrix for JAGS)}
\usage{
make_mat(dat, row = "rteno", col = "gridcellid", val, replace.na = FALSE)
}
\arguments{
\item{dat}{Data frame}

\item{row}{Variable in dat containing target row names}

\item{col}{Variable in dat containig target column names}

\item{val}{Variable containing the target cell contents}

\item{replace.na}{Whether to replace NA values with zero.}
}
\description{
Make 2-D array (Matrix for JAGS)
}
