% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{make_raster}
\alias{make_raster}
\title{Create a Spatial Sampling Grid as Raster using WorldClim}
\usage{
make_raster(
  baselayer = "worldclim",
  dir.out = NULL,
  resolution = 10,
  states = NULL,
  restrict.uscan = TRUE,
  buffer.dd = 5,
  countries = c("US", "CA")
)
}
\description{
Create a Spatial Sampling Grid as Raster using WorldClim
}
