% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_grid.R
\name{make_spatial_grid}
\alias{make_spatial_grid}
\title{Create a Spatial Sampling Grid}
\usage{
make_spatial_grid(
  dir.out,
  overwrite = TRUE,
  states = NULL,
  countries = c("USA", "CAN", "CA", "United States", "Canada",
    "United States of America")
)
}
\arguments{
\item{dir.out}{Where to save the resulting grid (as .RDS).}

\item{overwrite}{TRUE will overwrite existing \code{grid.rds} in \code{dir.out}}

\item{states}{Vector of states to which the spatial grid will be clipped.}

\item{countries}{Vector of countries. Defaults to a base map of USA and CAN, unless arg \code{states} is provided.}
}
\description{
Create a Spatial Sampling Grid
}
