% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_grid.R
\name{make_spatial_grid}
\alias{make_spatial_grid}
\title{Create a Spatial Sampling Grid}
\usage{
make_spatial_grid(
  dir.out,
  overwrite = TRUE,
  countries = c("US", "CA"),
  states = NULL,
  crs.target = 4326,
  hexagonal = TRUE,
  grid.size = 1,
  conterminousUSA = TRUE,
  ...
)
}
\arguments{
\item{dir.out}{Where to save the resulting grid (as .RDS).}

\item{overwrite}{TRUE will overwrite existing \code{grid.rds} in \code{dir.out}}

\item{countries}{Vector of countries. Defaults to a base map of USA and CAN, unless arg \code{states} is provided. If arg \code{states} is provided, this argument will be ignored. Must be specified using ISO-A2}

\item{states}{Vector of states to which the spatial grid will be clipped. Must be specified using ISO 3166-2 (see \url{https://en.wikipedia.org/wiki/ISO_3166-2})}

\item{crs.target}{Target CRS number for spatial grid.}

\item{hexagonal}{logical if TRUE will produce a spatial grid with hexagonal, as opposed to rectangular, cells}

\item{grid.size}{numeric size (relative to units defining crs.target) of resulting cell. E.g., if crs.target==4326 a value of gridsize=1.0 equals ~111.11km}

\item{conterminousUSA}{logical if TRUE will remove Hawaii and Alaska. Default TRUE to reduce computation times and extent of sptaial grid.}

\item{...}{Additional arguments}
}
\description{
Create a Spatial Sampling Grid
}
