% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_col_names.R
\name{match_col_names}
\alias{match_col_names}
\title{Munge eBird and BBS Column Names}
\usage{
match_col_names(x)
}
\arguments{
\item{x}{a data frame or matrix of eBird or BBS observations. If no column names match function target, will return same object.}
}
\description{
Ensure consistency among eBird and BBS datasets for integration
}
\keyword{internal}
