% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_ebird.R
\name{munge_ebird}
\alias{munge_ebird}
\title{Munge eBird Data}
\usage{
munge_ebird(
  fns.obs,
  fns.samps,
  dir.out,
  zerofill = TRUE,
  years = NULL,
  countries = c("US", "CA"),
  states = NULL,
  species = NULL,
  protocol = c("Traveling", "Stationary"),
  remove.bbs.obs = TRUE,
  max.effort.km = NULL,
  max.effort.mins = NULL,
  max.birds.checklist = NULL,
  max.num.observers = 10,
  complete.only = TRUE,
  ncores = NULL,
  ydays = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{fns.obs}{Filenames to observations files}

\item{fns.samps}{Filenames to sampling events file}

\item{dir.out}{Directory of where to save the filtered and munged data files}

\item{zerofill}{....}

\item{years}{....}

\item{countries}{....}

\item{states}{....}

\item{species}{....}

\item{protocol}{....}

\item{remove.bbs.obs}{....}

\item{max.effort.km}{....}

\item{max.effort.mins}{....}

\item{max.birds.checklist}{....}

\item{max.num.observers}{....}

\item{complete.only}{....}

\item{ncores}{....}

\item{ydays}{....}

\item{overwrite}{....}
}
\description{
Filter and zero-fill ebird data.
}
