% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_ebird_data.R
\name{munge_ebird_data}
\alias{munge_ebird_data}
\title{Create and Write or Load In the Filtered eBird Data}
\usage{
munge_ebird_data(
  fns.ebird,
  dir.ebird.out,
  overwrite = FALSE,
  countries = NULL,
  states = NULL,
  complete.only = TRUE,
  protocol = c("Traveling", "Stationary"),
  species = "Double-crested Cormorant",
  remove.bbs.obs = TRUE,
  years = NULL,
  max.effort.km = NULL,
  max.effort.mins = NULL,
  max.num.observers = 10,
  f_obs_out = paste0(dir.ebird.out, "ebird_obs_filtered.txt"),
  f_samp_out = paste0(dir.ebird.out, "ebird_samp_filtered.txt")
)
}
\arguments{
\item{fns.ebird}{File paths for the EBD and SamplingEvents data to import. Character vector of filenames for original files.}

\item{dir.ebird.out}{Location of where to save and/or find the filtered/subsetted eBird data.}

\item{overwrite}{logical If true will overwrite existing filtered data objects in directory \code{dir.ebird.out}}

\item{countries}{...}

\item{states}{...}

\item{complete.only}{logical if TRUE will remove incomplete checklists}

\item{protocol}{which eBird protocol to include}

\item{species}{species for filtering eBird data}

\item{remove.bbs.obs}{logical if TRUe will use crude and untested method for removing BBS data from eBird data (see function for detials)}

\item{years}{years to include in data}

\item{max.effort.km}{maximum distance (kilometers) of birding events to include}

\item{max.effort.mins}{maximum number of minutes of birding events to include}

\item{max.num.observers}{maximum number of observers delcared at a birding event to include}

\item{f_obs_out}{Filename for saving the filtered eBird observations data}

\item{f_samp_out}{Filename for saving the filtered eBird sampling data}
}
\description{
Filter the eBird data and sampling events using R package AUK.
}
