% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj.shorthand}
\alias{proj.shorthand}
\title{Auto generate a subdirectory name based on project parameters}
\usage{
\method{proj}{shorthand}(species, regions, grid.size, year.range)
}
\arguments{
\item{species}{string of species names using ISO-366-2}

\item{regions}{string of countries or states}

\item{year.range}{Vector of years. will take the min and max value}
}
\description{
Auto generate a subdirectory name based on project parameters
}
