% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC_allcode.R
\name{run_MCMC_allcode}
\alias{run_MCMC_allcode}
\title{Run MCMC in or Out of Parallel in nimble}
\usage{
run_MCMC_allcode(seed = 1, data, model, inits, mcmc.specs, verbose = FALSE)
}
\arguments{
\item{seed}{for internal use}

\item{data}{data and constants as list}

\item{model}{filepath or nimble model object}

\item{inits}{list of initial values}

\item{mcmc.specs}{MCMC specifications as a list; created using \code{set_mcmc_specs}}

\item{verbose}{logical if TRUE and code is not running in parallel will print all output available from Nimble commands.}
}
\description{
Run MCMC in or Out of Parallel in nimble
}
