% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC_allcode.R
\name{run_MCMC_allcode}
\alias{run_MCMC_allcode}
\title{Run MCMC in or Out of Parallel in nimble}
\usage{
run_MCMC_allcode(seed = 1, data, model)
}
\arguments{
\item{seed}{for internal use}

\item{data}{data and constants as list}

\item{model}{filepath or nimble model object}
}
\description{
Run MCMC in or Out of Parallel in nimble
}
