% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-saveruntime.r
\name{save.runtime}
\alias{save.runtime}
\title{Write Model Runtimes to Working Directory}
\usage{
save.runtime(name = NULL, time = NULL, other = NULL, fn = "model-runtimes.txt")
}
\arguments{
\item{name}{description of model to be saved under heading 'name'}

\item{time}{optional runtime as numeric or integer}

\item{other}{optional information to be saved under heading 'other' (e.g., number iterations, number chains, etc.)}

\item{fn}{path to existing or desired output file. This is where the runtimes and arguments will be saved. If file DNE, this function will create it.}
}
\description{
Write Model Runtimes to Working Directory
}
