% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_files.R
\name{scan_files}
\alias{scan_files}
\title{Check For Existing Files}
\usage{
scan_files(
  dir.proj,
  scan.for = c("jdat", "jags", "ebird_filtered", "bbs_obs", "grid", "bbs_spatial",
    "ebird_spatial", "ebird_zf")
)
}
\arguments{
\item{dir.proj}{project directory, within which first level subdirectories may include c(jags, ebird, bbs, spatial)}

\item{scan.for}{which types of files to scan for. Must be one or more of c("jags","ebird_filtered", "bbs_obs", "grid", "bbs_spatial", "ebird_spatial", "ebird_zf"). If multiples are provided and found, the returned object will be a named list of those imported files.}
}
\description{
Checks for existing files to import to avoid length data munging processes.
}
