% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mcmc_specs.R
\name{set_mcmc_specs}
\alias{set_mcmc_specs}
\title{Set Specifications for MCMC}
\usage{
set_mcmc_specs(
  na = 1000,
  nb = 5000,
  nc = 3,
  ni = 20000,
  nt = 10,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{na}{number of adaptation iterations}

\item{nb}{number of burn-in iterations per chain}

\item{nc}{number of chains}

\item{ni}{number of iterations per chain}

\item{nt}{thinning rate (every nth iteration removed)}

\item{ncores}{number of cores to run model in parallel}
}
\description{
Set Specifications for MCMC
}
