% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-trySaveModel.R
\name{trySaveModel}
\alias{trySaveModel}
\title{Ensure JAGS Model Outputs are Saved
Sometimes when working in OneDrive/Dropbox or on company networks, you may have problems saving to file. Running this instead of directly saving will ensure model output (.rds) is saved to an alternate location. This may be handy for running models in parallel/loops.
The function returns the output filename. Will save to a directory within the user's Desktop (windows, mac) if alt.save.dir is unspecified}
\usage{
trySaveModel(
  obj,
  obj.name = "mymodel",
  save.dir,
  alt.save.dir = paste0(file.path(Sys.getenv("USERPROFILE"), "Desktop"),
    "/JAGS.outputs/")
)
}
\arguments{
\item{obj}{an R object to save to file as .rds. Under the this context, a JAGS or other model output object.}

\item{obj.name}{desired name of the object file name}

\item{save.dir}{directory location of where to attempt to save the output .rds file}

\item{alt.save.dir}{if object save fails to save to \code{save.dir}, will save on user's desktop (future package edits will save to package files)}
}
\description{
Ensure JAGS Model Outputs are Saved
Sometimes when working in OneDrive/Dropbox or on company networks, you may have problems saving to file. Running this instead of directly saving will ensure model output (.rds) is saved to an alternate location. This may be handy for running models in parallel/loops.
The function returns the output filename. Will save to a directory within the user's Desktop (windows, mac) if alt.save.dir is unspecified
}
