% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-zerofill_ebird.R
\name{zerofill_ebird}
\alias{zerofill_ebird}
\title{Create a Zero-filled Data Object for eBird Observations}
\usage{
zerofill_ebird(
  list,
  cols.remove = c("SUBSPECIES_COMMON NAME", "TAXONOMIC_ORDER", "LAST_EDITED_DATE",
    "CATEGORY", "APPROVED", "REVIEWED", "SPECIES_COMMENTS", "...48", "HAS_MEDIA",
    "REASON", "TRIP_COMMENTS")
)
}
\arguments{
\item{list}{A list containing two named data frames, c("observations", "sampling"). This object is the result of \code{filter_ebird_data()}.}

\item{cols.remove}{A vector of column names to be excluded from the output file.}
}
\description{
Creates a zero-filled data object comprising the eBird observations and sampling events data supplied.
}
